Added the setup.sh script to set up all scripts.
[Details]
Now on any new computer, just clone the repo and run `chmod 700
setup.sh`, and then `./setup.sh`. Once it is done, it will ask you to
reload your terminal to source the files.

[Note]
The script just creates soft links for each file to the home directory,
so that the scripts are accessible by the necessary programs.

[File's changed]

	modified:   .gitconfig
	new file:   setup.sh