Updated pdftex to handle main files.
[Details]
Tex has an import feature where a main file can be specified and other
files imported from this. The previous pdftex could not handle this from
within other applications like vim, and hence the ability to specify a
main tex file for a particular directory was added.

[Note]
Minor changes to the vimrc were added.

[File's changed]

	modified:   .vimrc
	modified:   bin/pdftex