Added Reorder function.
[Details]
For the most part when working in latex, it is often the case that
something needs to be added within a particular paragraph. This presents
a problem for the style of the file in vim where the column width is set
to be 80 characters, and so many merges and line breakings need to
happen to get the source code looking acceptable again. The Reorder
function solves this by doing this process automatically.

[Note]
The Reorder command has been mapped to R (capital 'r').

[File's changed]

	modified:   .vimrc