Fixed bug and implemented better stop for Reorder
[Details]
Implemented a stopping criteria for the Reorder function where the
function does not try to merge a blank line (line with only spaces), and
instead assumes that to be the end of the paragraph. This is acceptable
as merging a blank line would not change the length of the line anyway,
and so the reorder script would only be removing blank lines
unnecesarily.

[Bugs]
Fixed a bug with the break command that broke words up. This happened
when the word did not end with a space and because the command expected
the space and moved backwards one character.

[File's changed]

	modified:   .vimrc