Updated reorder function.
[Details]
The reorder function in vim was exhibiting some strange behaviour due to
the way the 'e' and 'w' commands are implemented in vim. To alleviate
this, the commands were not used and instead, a customized function
Break() was implemented. This allows the user to break the current line
on the next available space, even if the space is currently at the
cursor. Using this function allows the reorder function to break lines
without punctuation going to the next line, or lines breaking at
improper places.

[Bugs]
The Reorder function is still not perfect in the sense that sometimes we
want small words to be able to go over the 80 character limit, and
sometimes we want big words not to. Getting a solution that adhears to
this fine line will be difficult though.

[File's changed]

	modified:   .vimrc