Updated pdftex script.
[Details]
Added texfot to the pdftex script. This allows the output of pdflatex to
be suppressed apart from warnings and errors. Also made the pdflatex
scripts run in non-stop mode, which runs passed errors (if it can) and
generates the pdf anyway. Finally, some heading lines were added, and
the colour of these lines changed so that the process can be easily
identified.

[File's changed]

	modified:   pdftex