Updated the tmx scripts to have saved formats.
[Details]
Added the feature to save a tmux format using the new addTmuxFormat bash
function. Once saved, the format appears in the bash completion list and
can be loaded by using the same name it was saved with in the tmx
script.

[Note]
This commit unfortunately also adds some pathing in the bash script,
this will have to be dealt with on the other sides...

[File's changed]

	modified:   .bashrc
	modified:   bin/tmx
	modified:   completion.bash