Separated out bash local and global settings.
[Details]
Moved things in .bashrc that are shared among set-ups to the .bash_global,
which can be distributed with this repository. Local settings such as
directory names are now contained in the normal .bashrc, which sources
the .bash_global to load global settings. Also adjusted the setup script
to deal with this change.

[Bugs]
Changed setup script to not copy the README.

[File's changed]

	renamed:    .bashrc -> .bash_global
	modified:   setup.sh