Fixed line merging for merge and reorder.
[Details]
Fixed line merging (which is used in the reordering) by adding the
Merge function. The default vim command to merge lines adds two spaces
when merging a line ending in a full stop. To rectify this, the newly
added merge function merges normally and removes the extra space if the
extra space was added.

[File's changed]

	modified:   .vimrc