Added YCM popup scrolling.
[Details]
The vim youcompleteme plugin gives a nice popup with the documentation
for the element under the cursor. Unfortunately, the popup only supports
mouse click events natively to do things like scroll the popup. This then
doesn't work in the setup of vim I am using. To remedy this, this commit
creates a function to get the popup window and scroll it, and some remaps
for ease of access.

[File's changed]

	modified:   .vimrc