Fixed up tmx command and all bash history.
[Details]
Added improvements to the tmx command so that it respects the bash
history of the shell that opens it, and adds the history of the tmux
session back to that shell on exit. To do this history editing, the
script needed to be sourced which is now done as a function in the
bash_global. Also added options to the bash_global for managing bash
history more naturally in general.

[File's changed]
	modified:   .bash_global
	modified:   bin/tmx